conformation avant de modifier ou supprimer

+++ app/src/main/java/com/example/gestiondecontacts/ContactDetailActivity.java
@@ -8,6 +8,7 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.ViewModelProvider;
 
@@ -34,9 +35,18 @@ public class ContactDetailActivity extends AppCompatActivity {
         deleteButton.setOnClickListener(view -> {
             long contactId = getIntent().getLongExtra("CONTACT_ID", -1);
             if (contactId != -1) {
-                contactViewModel.deleteContactById(contactId);
-                Toast.makeText(ContactDetailActivity.this, "Contact supprimé", Toast.LENGTH_SHORT).show();
-                finish();
+                new AlertDialog.Builder(ContactDetailActivity.this)
+                        .setTitle("Supprimer Contact")
+                        .setMessage("Êtes-vous sûr de vouloir supprimer ce contact ?")
+                        .setPositiveButton(android.R.string.yes, (dialog, which) -> {
+                            // Supprimez le contact de la base de données
+                            contactViewModel.deleteContactById(contactId);
+                            Toast.makeText(ContactDetailActivity.this, "Contact supprimé", Toast.LENGTH_SHORT).show();
+                            finish();
+                        })
+                        .setNegativeButton(android.R.string.no, null)
+                        .setIcon(android.R.drawable.ic_dialog_alert)
+                        .show();
             } else {
                 Toast.makeText(ContactDetailActivity.this, "Erreur lors de la suppression du contact", Toast.LENGTH_SHORT).show();
             }
